+++ app/plots/intervalplot.h
@@ -14,7 +14,7 @@ public:
     //~IntervalPlot();
     bool event(QEvent *event) override;
     void retranslate();
-    void scaleFont(double scaleFactor) override;
+    void scaleFont(float scaleFactor) override;
 
     void setup(QPair<int, int> points, QColor color);
 private:
